refactor(opentofu-agent): pattern conformance from review feedback
- Rename state.go -> agent_state.go (kosmoz)
- Move OpenTofuState struct from internal/db -> internal/types (kosmoz)
- Use r.PathValue() instead of chi.URLParam() (kosmoz)
- lockInfo struct -> map[string]string (forward compat with OpenTofu
  HTTP backend protocol; kosmoz)
- Long line wrap in array_agg query
